--- src/main/java/com/example/hellowtalk/core/dto/response/LoginResponse.java
@@ -1,4 +1,0 @@
-package com.example.hellowtalk.core.dto.response;
-
-public record LoginResponse(String accessToken) {
-}
