+++ src/main/java/com/example/hellowtalk/core/user/controller/HelloController.java
@@ -1,4 +1,4 @@
-package com.example.hellowtalk.core.controller;
+package com.example.hellowtalk.core.user.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
