+++ src/main/java/com/example/hellowtalk/core/user/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.example.hellowtalk.core.dto.request;
+package com.example.hellowtalk.core.user.dto.request;
 
 public record LoginRequest(String username, String password) {
 }
