+++ src/main/java/com/example/hellowtalk/core/user/dto/request/UserCreateRequest.java
@@ -1,4 +1,4 @@
-package com.example.hellowtalk.core.dto.request;
+package com.example.hellowtalk.core.user.dto.request;
 
 public record UserCreateRequest(String username, String password, String name) {
 }
