+++ src/main/java/com/example/hellowtalk/core/user/dto/response/LoginResponse.java
@@ -1,0 +1,4 @@
+package com.example.hellowtalk.core.user.dto.response;
+
+public record LoginResponse(String accessToken) {
+}
