+++ src/main/java/com/example/hellowtalk/core/user/service/AuthService.java
@@ -1,8 +1,8 @@
-package com.example.hellowtalk.core.service;
+package com.example.hellowtalk.core.user.service;
 
-import com.example.hellowtalk.core.dto.request.LoginRequest;
-import com.example.hellowtalk.core.entity.User;
-import com.example.hellowtalk.core.repository.UserRepository;
+import com.example.hellowtalk.core.user.dto.request.LoginRequest;
+import com.example.hellowtalk.core.user.entity.User;
+import com.example.hellowtalk.core.user.repository.UserRepository;
 import com.example.hellowtalk.global.auth.JwtProvider;
 import com.example.hellowtalk.global.exception.CustomException;
 import com.example.hellowtalk.global.exception.ErrorCode;
